requirements.txt
flask==2.3.3
werkzeug==2.3.7
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
tensorflow-cpu==2.7.0
protobuf==3.20.0
gunicorn==20.1.0
matplotlib==3.4.3
seaborn==0.11.2 